ARMV7: Fix booting when no relocation is needed
If the u-boot image is linked to already be at the end of the
SDRAM address range and relocation isn't needed, it would still not
run right because the dyn fixups weren't being done.  Also, it would
crash because r9 wasn't initialized when it jumped to the code to
find build_init_r().

Change-Id: I0fc2cdc6acd7be1e6d510b3f027f9f2379fcfd3d
Signed-off-by: Mike J. Chen <mjchen@google.com>