CMD_TERMINAL: Improve transparency, help and configuration
The terminal command was not very transparent.  Whenever a tilde
was sent, it would announce that it saw the tilde which could
corrupt the communication if you were trying to use the terminal
command to implement a pass-through.  In addition, a delay was
not required before the '~.' escape sequence so the likelihood of
an "accidental" detection of the escape sequence in the midst of
a transfer was higher.

The changes:
1) Added a guard delay that must be present before the escape
   sequence is detected.  In addition, there must be less than
   that delay between the two characters of the escape sequence.
2) Removed the announcing of noticing the escape characters.
3) The escape characters and the delay are configurable via
   #defines in the board's configuration header.
4) Improved the U_BOOT_CMD declaration to have the correct number
   of maximum arguments, not allow the command to be repeated
   with just an enter, and add a description of the required
   command line argument.

Change-Id: I84191bdf239236653cd2922484416251de4a0aba
Signed-off-by: Scott Anderson <saa@google.com>