FASTBOOT: Add config to prevent overwriting of bootargs
To allow customization of the kernel command line, add
CONFIG_FASTBOOT_PRESERVE_BOOTARGS which simply makes the header's
command line available in the hdr_cmdline environment variable
instead of always overriding bootargs with the header's command
line and the serial number.  This can be subsequently used when
bootargs is constructed.

Change-Id: I08ac381796bd2286feb97605648e21ae78338dd2
Signed-off-by: Scott Anderson <saa@google.com>