SERIAL: Add raw putc's that do not do LF->CRLF translations
The terminal command was changing each newline into a
carriage-return and a newline.  This was caused by
_serial_putc. There already is a _serial_putc_raw, however is was
not exposed via the serial_device nor stdio_dev structures so it
was not available.

I copied the putc interfaces with putc_raw versions and changed
the terminal command to use them.  The changes have been tested
and the CRLF "corruption" no longer occurs.

Change-Id: I9c291c5f4f9eda8a4674bb3524a9ed950fc2a3fc
Signed-off-by: Scott Anderson <saa@google.com>