I2C: omap: I2C read should not wait for bus idle after addr write
Commit 5bd6d1e9dd9097c0e86cb1549a4b6f4fec683544 introduced a
change to I2C reads that waits for the bus to be not busy after
writing the address, but before doing the read.  However, the bus
is still "busy" because the ack from the write will still be
holding down SDA.  This causes i2c_read_internal to complain
	"timed out in wait_for_bb: I2C_STAT=1004"
after it times out waiting for the bus to be idle using
wait_for_bb().

This change is to wait for the status change instead of the bus
to be idle.

Change-Id: Ib6d83cc340761f21d8c0216ee557c13156e85c13
Signed-off-by: Scott Anderson <saa@google.com>