Turn on SPL build.
SPL is TI's term for building the first stage bootloader from the
same code base as u-boot.  It's output is the MLO file that the xloader
tree used to build.

I've tested that the MLO output works on Tungsten so we can now
deprecate the xloader branch/project.

The build creats both the MLO and u-boot.bin file.  Only tricky part
was getting the mmc sector right for loading the u-boot image.

Also, the u-boot run address has been moved so anyone using usbboot needs
to update usbboot to match.  The new address is 0x80100000.

Change-Id: I9633b386abda5c3ea67a22fa92e4e6ceba634897
Signed-off-by: Mike J. Chen <mjchen@google.com>