ARMV7: OMAP4: Reduce power consumption in bootloader
On OMAP4460, changing the MPU to 1820MHz and enabling a number
of non-essential clocks causes the SOC to consume a lot of
power (always increasing due to the viscious cycle of heating
causing increased leakage current causing more heating and
more current).  We were setting 700+mA starting point, and
ramping up quickly.

With these changes, we start at closer to 550mA and still
ramp up, but slowly and seems to stabilize around the 640-650mA
range (still testing).

Not clear if there are some side-effects to this change in the
kernel.  I see HDMI booted and adb worked.

Change-Id: Ib105cdad4d036b8a5fc1d207f940f29c0f98cbae
Signed-off-by: Mike J. Chen <mjchen@google.com>