Use sysclk for the ABE instead of 32KHz
Make sure the tungsten bootloaders choose the sysclk as the source for the ABE
DPLL instead of the 32KHz clock.  The sysclk is the clk that we can slew using
the VCXO, so it is essential that the ABE sources from it in order to be able to
perform clock recovery and maintain audio sync with other tungstens.

Signed-off-by: John Grossman <johngro@google.com>
Change-Id: I5faaef632c1f42901bc9fe106e2fe16d10e05e5a