Enable I2C4 on OMAP4 in uboot.
OMAP4 has 4 I2C controllers, not 3.  (technically, it has 5, but only 4 in the
L4_PER domain).  Add the definition for base address of the I2C4 unit, raise the
limit for OMAP4 and tweak the OMAP[234] i2c_set_bus_num to allow us to choose
I2C4 on OMAP4.

Signed-off-by: John Grossman <johngro@google.com>
Change-Id: I96b525df1c7e455ade5926432e874504d1b75b92