Don't change the state of unused pins in SPL.
SPL only needs to be certain of the state of 2 sets of pins; the MMC1 controller
and the serial port.  In the pursit of leaving the pin muxes at their minimal
configuration, I am shutting off the configuration of MMC2, the CTS/RTS lines on
UART3, and the various I2C busses.

Signed-off-by: John Grossman <johngro@google.com>
Change-Id: Ia319ce1cedcab6aac88fb45902a47a987b2883a3