ARMV7: tungsten: Add board_fbt_start() and board_fbt_end()
Set a LED pattern when in fastboot mode.  I think spec wants us to
actually animate a pattern but I don't know the pattern yet and this
is easier and lower power for now.

Also add call to vcnl_get_promixity() in board_fbt_key_pressed().
The fully enclosed system is not giving reliable values yet but
I put in some initial values for an uncovered system.

Change-Id: I1a54c0884370ff494365b00ae4b7ebca110b6f03
Signed-off-by: Mike J. Chen <mjchen@google.com>