mkimage: Fix an off by 8 error in gp_header
The ROM bootloader appears to treat the size in the gp_header to be
relative to the start of the gp_header, but the code was just passing
the size of the bin file following the gp_header.  The ROM bootloader
would also round this value up to a MMC page size, so the mismatch would
not be apparent unless the bin file size was exactly a page size multiple.
Now add the gp_header size itself into the size stored.

Change-Id: Idaebc27500406346776420c03fd192ee9cf20a08
Signed-off-by: Mike J. Chen <mjchen@google.com>