Makefile: Define CONFIG_MFG for cpp as well as cc
When the dependencies are made, CPPFLAGS is used instead of
CFLAGS (see rules.mk).  We were adding CONFIG_MFG only to CFLAGS
before and that could cause the making of dependencies to fail
because required defines that are MFG dependent may not be
defined.  The example I ran into was that I was only defining
CONFIG_SMC911X_32_BIT for MFG, but then when the dependencies
were generated for MFG, drivers/net/smc911x.h would throw an
error because CONFIG_SMC911X_32_BIT was not defined.

Change-Id: I02dd9f67be464ae9344c9bcdd6b875366b731b75
Signed-off-by: Scott Anderson <saa@google.com>