Generate missing MAC addresses from Die ID.
When the various mac addresses needed for tungsten are missing from the device
info partition, generate them using a pseudorandom number generator seeded with
the Die ID and a per-MAC-address-type index.  This is mostly for manufacturing
diagnostics where we need addresses to run diagnostics, but don't want to flash
the device_info partition until the very last second.  The added bonus of this
is that un-initialized tungstens will have some form of addresses which stay
static across boots instead of getting the kernel's randomly generated
addresses.

Signed-off-by: John Grossman <johngro@google.com>
Change-Id: I98220acfb71037c31e1e2e1e6388d553f4c9de73