ARMV7: tungsten: Change to lower MPU operating frequency and voltage
In our ongoing battle to keep power and heat down in the bootloader,
set the MPU operating frequency (it's actually an increase from
our previous bypass mode of 38.4MHz to 349.6MHz) and TPS voltage
to 1025mV (from default of 1430mV).  This drops the measured
power consumption when sitting in fastboot from about 550mA to about
350mA.  This makes the enclosed system boards run stabling in
fastboot for much longer (indefinite?  time will tell).

Change-Id: Ib30547526a7240ad71ca9017559e866a5332e560
Signed-off-by: Mike J. Chen <mjchen@google.com>