Don't dump default MACs in MFG builds.
Instead of dumping default MAC addresses every time to the serial
console in MFG builds, Celestica would prefer to fetch the data using
the diagnostic console and the printenv command.

This change makes the MFG build set environment variable which can be
fetched with printenv instead of just dumping to the serial console
every time.