mmc: Change erase to do large contiguous blocks in one cmd
It's a lot faster when erasing many contiguous blocks to issue
one command instead of doing it one erase group at a time because
the MMC part can often do it in parallel.

The issue it wasn't done earlier is timeouts.  Specific to omap_hsmmc,
make the erase code check for completion status after the erase
command has been sent, and adjust the timeout used in that case
to be relative to the number of blocks being erased.  Other
mmc drivers might need adjusting too to match this behavior and
the ratio of timeout to number of blocks being erased is really
device specific and not driver specific but for now we use a
fixed ration of 1 second per 1000 blocks.

Change-Id: I5e25c64d475ad8b5508926347e480ddcb7e5bb05
Signed-off-by: Mike J. Chen <mjchen@google.com>