MTD NAND: Fix NAND YAFFS OOB handling.
The handling of the WITH_YAFFS_OOB flag in nand_write_skip_bad()
was flawed in two ways:

1) The code that does the nand_write before the while loop can
not be used in the case of WITH_YAFFS_OOB because it does not
write the OOB data.  I fixed the flags check to skip the code if
WITH_YAFFS_OOB was set.

2) After #1 was fixed, it was observed that only the first block
was written.  The flaw is that the sense of the check for the
return status of nand->write_oob was reversed.  It would break
out of the loop if it succeeded so only the first block was
written.

Change-Id: I89a5f6afd698ae74da9bdc136e81e3c00a21eaa5
Signed-off-by: Scott Anderson <saa@google.com>