ARMV7: omap4: clocks: allow build flag to force use of TPS62361
For testing/development, sometimes a OMAP4430 is put in a board
that has a TPS62361 (i.e. a board designed to support the OMAP4460).
The old runtime checks weren't enough to handle that case so create
a compile time flag that can bypass the runtime checks and force
the usage of the TPS62361.

Change-Id: I6f663a591e9e94a2559997ebab00fa818bcd15fb
Signed-off-by: Mike J. Chen <mjchen@google.com>