FASTBOOT: Add 'fastboot oem erase <partition> <blk_cnt>' support
This is similar to 'fastboot erase <partition>' except that
it allows for an optional blk_cnt argument which is used
to limit the amount of erasing to be less than the full
partition size.  This is especially useful to do a quick
partial erase of userdata.  Enough to have setup_fs detect
it needs to be formatted, but faster than erasing the
really huge userdata partitions we see often times.

Change-Id: I786b58bece0606fabc5e0fdfa17bd852b6a29ad4
Signed-off-by: Mike J. Chen <mjchen@google.com>