CMD_BLK: Enhance with new options
With the new partition access functions, adding the ability to
specify a partition instead of an offset became easy.  I also
allow the count to be not specified if the partition was given
and it will assume the entire partition.  Unfortunately, because
of the changes in parsing, the code was heavily restructured.

Change-Id: I45b49b8879bf4c07a01f01dc83b7e857cd2816e7
Signed-off-by: Scott Anderson <saa@google.com>