Don't stop in fastboot based on the prox sensor.
Need to revisit this later.  Right now, using the proximity
sensor as an absolute trigger to force into fastboot is causing all
sorts of problem with false positive triggering, particularly in
non-benchtop tungsten units.  #if 0 this out for now until a final
decision has been made about what the HW solution will be (cap-sense
vs. IR prox vs. rotary encoder) and what the SW flow will be as a
result of this.

Change-Id: I4c6317efcd4f4b222da3c78d770a15786a08b077
Signed-off-by: John Grossman <johngro@google.com>