Diags: change the way we force fastboot in MFG builds.
Previously, we were forcing fastboot in mfg builds by playing games
with the preboot and bootcmd config options.  Unfortunately, the
results of these games were getting saved in the env partition which
was having unintended side-effects.  Instead of continuing to hack on
the old approach, this change switches to an approach where we just
have a top-level config override which is interpreted by main.c as it
is selecting a bootcmd from its various sources.

Change-Id: I442d7ee003ee3c674ac8bc1a546a1d10b56323dd
Signed-off-by: John Grossman <johngro@google.com>