FASTBOOT: Clean up return handling
The fastboot code had a number of functions that always returned
zero and/or whose return values were always ignored.  To me, this
made working on the code harder as it was unclear when functions
were called, what to do with return values.  I went through and
changed all of functions that returned useless or unused int
values to be void.

There were a few places where return values were being ignored,
but added some diagnostics based upon the return value.

Finally, in do_booti(), the handling of the return was
inconsistent and memory could be leaked due to not calling
free().

Change-Id: I3778d0723699e31f46c061a81df44d7b16a5706a
Signed-off-by: Scott Anderson <saa@google.com>