MMC: Make sure PBIAS config matches VMMC supply.
Depending on the system, VMMC may end up being 3 or 1.8v.  It is
important that the PBIAS cell be configured to match the supply.
Fortunately, the PBIAS cell contains a handy detector which can
determine what the supply is set to after PWRDNZ has been released.
Use this detector to make sure that the cell is properly configured
and does not damage itself.

Change-Id: I96d8aff3d19f5ea783ae7eb646a7e3fa920a64be
Signed-off-by: John Grossman <johngro@google.com>