FASTBOOT: Switch from MMC to generic block device
At a high level, this change replaces the MMC specific code in
fastboot with generic block device code.  In finer detail, this
change includes:

1) The sequence find_mmc_device(), mmc_init(), mmc_get_dev() is
replaced with just get_dev_by_name() with the assumption that the
block device is initialized in the board specific startup code.

2) Instead of calling get_dev() in multiple places, change the
code to set priv.dev_desc in fbt_preboot() and then rely on that
value (with checks early in the commands to ensure it is set).

3) The fastboot specific partition code and data structures were
replaced with generics.  For example, import_efi_partition() in
fastboot.c is replaced with just calling the generic
get_partition_info().  This allows the board_fbt_load_ptbl() that
was in fastboot.c to be generalized enough that it can just serve
as the default board_fbt_load_ptbl in cmd_fastboot.c.  It is
still a weak alias, so if need be, it could be overwritten.

4) As much as possible, the code to read/write/erase was replaced
with calls to the partition_* functions to do the same thing.
Where that was not possible, calls to run the pre_/post_
partition commands stored in environment variables to allow
boards to set block devices up in a partition specific manner.

5) fbt_handle_erase() was demoted back to just handling the erase
command instead of serving as generic partition erase code.

6) In several places, the calculations involving blocks were
fixed to prevent overflow or to remove assumptions about 512
bytes/block.

7) The booti command was reworked.  It used to require a
parameter that specified "mmc0" or "mmc1".  In reality, the
partition table that the booti command used was from whichever
MMC was specified at compile time and trying to use that
partition table for a different device would probably have caused
disappointment.  It now can only take a single optional parameter
that specifies either an address or a partition name.  I also
cleaned up booti's help text to be formatted correctly.

Change-Id: If8e2a61900c27f6c4df609efdf36895f6f16bf1e
Signed-off-by: Scott Anderson <saa@google.com>