FASTBOOT: Clean up some DEBUG and INFO code
Changed code that used #ifdef DEBUG and #ifdef INFO to use the
already defined FBTDBG and FBTINFO macros.

Removed a #ifdef INFO printf(".\n") that was immediately followed
by a much more informative FBTINFO.

Fixed the formats on a couple of FBTINFOs to fix compiler
warnings.

Commented out an FBTDBG that I suspect caused infinite recursion
when sending to USB.

bootimg_print_image_hdr is a function that is used by do_booti()
to print out the fastboot image header if DEBUG is defined.  It
was called if an address was given to booti, but not when a
partition was given.  It now does it for both cases.

Change-Id: Iaff133c820a4b15c350b4b80c11412d052a266b1
Signed-off-by: Scott Anderson <saa@google.com>