omap: Add CONFIG_SPL_VERIFY_IMAGE support
This flag turns on image verification (crc in header and
data).  It'll only work if the u-boot.img file is used since
it requires the header to do the verification.

It also adds support for an alternate MMC location to find
a second copy of the bootloader if the first one fails to
verify.

Change-Id: Ibfdbfd91dea0c725d9ce9d081e3004833a91ba47
Signed-off-by: Mike J. Chen <mjchen@google.com>