ARMV7: tungsten: Add a second bootloader partition and turn on verification
To make OTAs more failsafe, we're adding a second bootloader partition
(name is 'bootloader2') and turning on image verification (does a crc
validation of the image).  This requires that we flash u-boot.img instead
of u-boot.bin, since the former has the header w/ crc information that we
need.

While I was in there, I moved the efs partition, which is relatively small,
next to the other small partitions.

Lastly, updated build_server.sh to make the u-boot.img file available
instead of the u-boot.bin file.

Change-Id: I1039fabff1a78f9c2c9be9ca690e64dfef7b7c20
Signed-off-by: Mike J. Chen <mjchen@google.com>