MMC: Fix, enhance and clean up interpretation of CID
The MMC Device Identification register (CID) is used to populate
the block device's product, vendor and revision strings.  The old
code did not match the standard in how the fields were
interpreted. In addition, not all of the fields were used.  In
this change, I fixed the interpretation of the fields, got all
the fields (other than the CRC) output, and in an attempt to make
the code clearer I separated the unpacking of the fields from the
formatting.

This information can be printed using the "blk dev" command and
is useful in getting detailed information about MMCs.

Change-Id: Ied89b7570ecd440ba586249c0aa178a7325e283b
Signed-off-by: Scott Anderson <saa@google.com>