ARMV7: OMAP: Fix roll over for get_timer_masked()
get_timer_masked() is the function that underlies get_timer() and
get_ticks().  It had the flaw that when the underlying hardware
timer would overflow (which happens every 14.9 minutes on the
platform I'm using), the returned value would jump from 894784
(the overflow value converted into milliseconds) back down to 0.
The prior code tried to handle that with:

  gd->tbl += ((TIMER_LOAD_VAL / (TIMER_CLOCK / CONFIG_SYS_HZ))
               - gd->lastinc) + now;

however, TIMER_LOAD_VAL is 0, so this turned into:

  gd->tbl += - gd->lastinc + now;

On the platform I was testing with, I was calling get_timer in a
loop.  When the hardware timer overflowed, gd->tbl and
gd->lastinc were both the maximum (894784) and now was 0 so this
code just reset the value to be returned back to zero.

The fixed code is:

  gd->tbl += ((TIMER_OVERFLOW_VAL / (TIMER_CLOCK / CONFIG_SYS_HZ))
               - gd->lastinc) + 1 + now;

TIMER_OVERFLOW_VAL is 0xffffffff, and on the platform I am using,
TIMER_CLOCK is (38400000 / (2 << 2)) and CONFIG_SYS_HZ is 1000,
so this turns into:

  gd->tbl += ((0xffffffff / ((38400000 / (2 << 2)) / 1000))
               - gd->lastinc) + 1 + now;

which is:

  gd->tbl += (894784 - gd->lastinc) + 1 + now;

For the roll over case where gd->tbl and gd->lastinc were both
the maximum and now was zero, this causes 1 to be added to
gd->tbl so 894785 is returned and the timer keeps going up until
it overflows the ulong.

I also fixed a couple typos in a nearby comment.

Change-Id: I6eb7eaeac974b4996f8a01860b8d9eb3708fef21
Signed-off-by: Scott Anderson <saa@google.com>