ARMV7: tungsten: Add board_fbt_handle_flash and support for tagged MLO file
Gives the bootloader the ability to handle a MLO file that has more
than one image, each signed for a different HS part, and the bootloader
can figure out which one to use for the current device.

The tag consists of a MPKH_info structure appended to the end of the
MLO file generated by the TI signing tools.  This structure has a four
character 'MPKH' tag followed by the 32 bytes of the MPKH that needs
to match the MPKH registers of the current device before we will flash
it.  For backwards compatibilty, we accept a non-tagged MLO file and
flash it blindly, assuming that the user knows what they're doing.

Change-Id: Ifa07e547472cef2a28fb63ced76f26acc817a199
Signed-off-by: Mike J. Chen <mjchen@google.com>