FASTBOOT: Make oem unlock_accept call recovery to do factory wipe
The bootloader already does an erase userdata before
changing the unlock state, but an erased userdata isn't formatted and
used to rely on setup_fs to do the formatting on the next Android
boot.  However, that is somewhat risky since the next boot becomes
longer and sometimes users remove power before setup_fs is done,
leaving a corrupted file system.

We now force a boot into recovery with --wipe_data argument, which
makes it do the format for us, and reduces the chance of a
bad userdata state.  The reboot into recovery will redo itself
if it doesn't finish and reboot cleanly.

Change-Id: I45a754793fe50d0cd3506a9782a95de5644c4842
Signed-off-by: Mike J. Chen <mjchen@google.com>