ARMV7: tungsten: Add support for more board revs
Since we're limited to 3 bits of board-id, we're going
to have to reuse the older ones.  Based on HW groups
current predictions of names, I've put in the values
for PVT, PROD, and PROD1 (if necessary) and use
a lookup table to convert from raw board-id values to
revision numbers.  We keep the revision numbers as
always increasing values to make it easy for code to
do comparisons.

Also removed code for disabling capsense in pre-DVT3
units since the AVR firmware does that itself now.

Change-Id: Ie879a35451464606f83db205020702c681cd0e55
Signed-off-by: Mike J. Chen <mjchen@google.com>