ARMV7: add DMA MMC writes to OMAP3 & OMAP4 devices
This commit enabled u-boot to do DMA writes to the MMC-based on the
CONFIG_OMAP_MMC_DMA_WRITE option. This is done using the
sDMA module of the OMAP, since that is the least common denominator between
OMAP3 and OMAP4, thus allowing OMAP3 to benefint from this commit too. Included
are: (1) include for SDMA addressed, structures, etc; (2) mini sDMA driver
(embedded in ompa_hsmmc.c for now); and (3) sDMA intergration for MMC writes.
Some code is present to demonstrate how to integrate read capability too, but
read functionality is *untested*

Change-Id: I74ce902fc4b88e783f2620188cff3ff27a7c1a4d