MD5 LIBRARY: Expose functions to do piecemeal MD5 calculation
The functions MD5Init, MD5Update and MD5Final are implemented in
lib/md5.c but they were static to that file.  This change
promotes them to the global name-space so they can be used in
code that wants to do an MD5 across more data than can fit in
memory at one time.

Change-Id: Ia9d73732268f6417ce267e4ea6dfb7e479ca1124
Signed-off-by: Scott Anderson <saa@android.com>