ARMV7: tungsten: define CONFIG_TWL6030_POWER
This makes usb work when we cold boot into fastboot mode
with no usb cable connected and then insert usb afterwards.
When usb cable is connected from the beginning, the ROM
bootloader was doing the TWL6030 setup, but if the cable
wasn't connected, the bootloader has to do the setup itself.

Change-Id: Iaa2ad0e19264699453e2b74575028f1b0fc59222
Signed-off-by: Mike J. Chen <mjchen@google.com>