ARMV7: OMAP4: USB: Configure twl6030x USB like ROM bootloader
This is a followup to commit 00261ba7a7e29c28708972c908a94993ea322dba.
That change worked in that USB worked in fastboot if the cable is
attached after boot, but had a bad side-effect that if we then
booted to OS (using fastboot continue or fastboot reboot), then
unplugged and replugged USB, USB wouldn't work.

I believe the VUSB_IN_VBAT was not the right setting, but to be
extra safe, I dumped the registers in question after the ROM
set them up in the cold boot with USB cable attached case, and
just made them the same.  CFG_LDO_PD2 and USB_VBUS_CTRL_SET
registers appeared to be already set to working values by
the ROM bootloader in all cases, and VBUS_IN_PMID is set instead
of VUSB_IN_VBAT.

This fixes the plug/unplug problem in adb.

Change-Id: I200e7701ca9905a6eba6f47ebd71787ea3dcb993
Signed-off-by: Mike J. Chen <mjchen@google.com>